feat: create post controller and implements WebApplicationException

+++ src/main/java/arg/acme/usecases/DeletarUserUseCase.java
@@ -5,6 +5,8 @@ import arg.acme.repository.UserRepository;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 
 import java.util.UUID;
 
@@ -17,6 +19,15 @@ public class DeletarUserUseCase {
 
     @Transactional
     public void deletarPessoa(UUID id) {
+
+        var user = userRepository.findById(id);
+        if (user == null) {
+            throw new WebApplicationException(
+                    Response.status(Response.Status.NOT_FOUND)
+                            .entity("User not found with id: " + id)
+                            .build()
+            );
+        }
         userRepository.deleteById(id);
     }
 }
